requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.2
backcall==0.2.0
backports.csv==1.0.7
beautifulsoup4==4.12.2
billiard==3.6.4.0
bs4==0.0.1
cached-property==1.5.2
camelot-py==0.11.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==39.0.2
cssselect==1.2.0
cssselect2==0.7.0
cv==1.0.0
cx-Oracle==8.3.0
dateparser==1.1.8
dateutils==0.6.12
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
discord==1.7.3
discord.py==2.0.1
distlib==0.3.6
distro==1.8.0
Django==4.1.5
django-auth-ldap==4.1.0
django-cors-headers==3.13.0
django-crequest==2018.5.11
django-debug-toolbar==4.1.0
django-drf-filepond==0.5.0
django-environ==0.9.0
django-filter==22.1
django-import-export==3.2.0
django-jazzmin==2.6.0
django-phonenumber-field==7.0.2
django-rest-framework==0.1.0
django-sequences==2.7
django-silk==5.0.4
django-simple-history==3.2.0
django-sslserver==0.22
django-storages==1.13.2
django-windowsauth==1.4.0
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0
djangorestframework-simplejwt==5.2.2
dnspython==2.4.2
docopt==0.6.2
et-xmlfile==1.1.0
exchangelib==5.1.0
executing==1.2.0
fake-useragent==1.3.0
feedparser==6.0.10
filelock==3.8.0
frozendict==2.3.8
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
gprof2dot==2022.7.29
greenlet==2.0.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
install==1.3.5
invoice2data==0.4.4
ipython==8.11.0
isodate==0.6.1
jedi==0.18.2
jugaad-data==0.24
kombu==5.2.4
ldap3==2.9.1
lxml==4.9.2
MarkupPy==1.14
matplotlib-inline==0.1.6
mod-wsgi==4.9.4
multidict==6.0.2
multitasking==0.0.11
nsetools==1.0.11
num2words==0.5.12
numpy==1.23.4
O365==2.0.31
oauthlib==3.2.2
odfpy==1.4.1
opencv-python==4.8.0.74
openpyxl==3.1.2
oscrypto==1.3.0
pandas==1.5.1
parse==1.19.1
parso==0.8.3
pbr==5.11.0
pdfkit==1.0.0
pdfminer.six==20221105
peewee==3.17.0
pexpect==4.8.0
phonenumbers==8.13.6
phonenumberslite==8.13.6
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.4
prompt-toolkit==3.0.38
psycopg2==2.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyee==8.2.2
pygame==2.5.2
Pygments==2.14.0
pyHanko==0.17.0
pyhanko-certvalidator==0.20.1
PyJWT==2.6.0
pypdf==3.5.1
pypng==0.20220715.0
pyppeteer==1.0.2
pyquery==2.0.0
pyspnego==0.10.2
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-ldap==3.4.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
regex==2023.6.3
reportlab==3.6.12
requests==2.31.0
requests-html==0.10.0
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
sgmllib3k==1.0.0
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
soupsieve==2.5
SQLAlchemy==1.4.16
sqlparse==0.4.3
sspilib==0.1.0
stack-data==0.6.2
stevedore==4.1.1
stringcase==1.2.0
svglib==1.5.1
tablib==3.4.0
tabula-py==2.7.0
tabulate==0.9.0
Tempita==0.5.2
tinycss2==1.2.1
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uritools==4.0.1
urllib3==1.26.12
use==0.1.0
vine==5.0.0
virtualenv==20.16.7
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
w3lib==2.1.2
wcwidth==0.2.6
webencodings==0.5.1
websockets==10.4
wfastcgi==3.0.0
xhtml2pdf==0.2.9
xlrd==2.0.1
xlwt==1.3.0
yahoo-fin==0.8.9.1
Yahoo-ticker-downloader==3.0.1
yarl==1.8.1
yfinance==0.2.31
zipp==3.17.0
